apply remaining minor feedbacks
